+++ webservice/src/main/java/edu/sjsu/cohort6/esp/service/rest/UserResource.java
@@ -17,6 +17,7 @@ package edu.sjsu.cohort6.esp.service.rest;
 import edu.sjsu.cohort6.esp.common.CommonUtils;
 import edu.sjsu.cohort6.esp.common.User;
 import edu.sjsu.cohort6.esp.dao.DBClient;
+import edu.sjsu.cohort6.esp.service.rest.exception.AuthorizationException;
 import edu.sjsu.cohort6.esp.service.rest.exception.BadRequestException;
 import edu.sjsu.cohort6.esp.service.rest.exception.InternalErrorException;
 import edu.sjsu.cohort6.esp.service.rest.exception.ResourceNotFoundException;
@@ -52,14 +53,18 @@ public class UserResource extends BaseResource<User> {
     @Consumes(MediaType.APPLICATION_JSON)
     public Response create(@Auth User user1, @Valid String userJson, @Context UriInfo info) {
         try {
-            User user = CommonUtils.convertJsonToObject(userJson, User.class);
-            List<User> userList = new ArrayList<>();
-            userList.add(user);
-            userDAO.add(userList);
-            URI uri = UriBuilder.fromResource(UserResource.class).build(user.getId());
-            return Response.created(uri)
-                    .entity(Entity.json(user))
-                    .build();
+            if (isAdminUser(user1)) {
+                User user = CommonUtils.convertJsonToObject(userJson, User.class);
+                List<User> userList = new ArrayList<>();
+                userList.add(user);
+                userDAO.add(userList);
+                URI uri = UriBuilder.fromResource(UserResource.class).build(user.getId());
+                return Response.created(uri)
+                        .entity(Entity.json(user))
+                        .build();
+            } else {
+                throw new AuthorizationException("User " + user1.getUserName() + " is not allowed to perform this operation");
+            }
         } catch (Exception e) {
             log.log(Level.SEVERE, "Error in adding user", e);
             throw new BadRequestException();
@@ -70,9 +75,13 @@ public class UserResource extends BaseResource<User> {
     @GET
     @Produces(MediaType.APPLICATION_JSON)
     public List list(@Auth User user, @QueryParam("filter") String filter) throws InternalErrorException {
-        List<String> userIds = new ArrayList<>();
-        List<User> userList = userDAO.fetchById(userIds);
-        return userList;
+        if (isAdminUser(user)) {
+            List<String> userIds = new ArrayList<>();
+            List<User> userList = userDAO.fetchById(userIds);
+            return userList;
+        } else {
+            throw new AuthorizationException("User " + user.getUserName() + " is not allowed to perform this operation");
+        }
     }
 
     @Override
@@ -80,32 +89,54 @@ public class UserResource extends BaseResource<User> {
     @Produces(MediaType.APPLICATION_JSON)
     @Path("/{id}")
     public User retrieve(@Auth User user, @PathParam("id") String id) throws ResourceNotFoundException, InternalErrorException {
-        List<String> userIdList = getListFromEntityId(id);
-        List<User> userList = userDAO.fetchById(userIdList);
-        if (userList != null && !userList.isEmpty()) {
-            return userList.get(0);
+        if (isAdminUser(user)) {
+            List<String> userIdList = getListFromEntityId(id);
+            List<User> userList = userDAO.fetchById(userIdList);
+            if (userList != null && !userList.isEmpty()) {
+                return userList.get(0);
+            } else {
+                throw new ResourceNotFoundException();
+            }
         } else {
-            throw new ResourceNotFoundException();
+            throw new AuthorizationException("User " + user.getUserName() + " is not allowed to perform this operation");
         }
     }
 
+    /**
+     * Update any user.
+     * This method is only accessible to admin user.
+     * Partial update of user field is not supported yet.
+     * Student user can use students endpoint to update his/her user info which supports partial updates.
+     *
+     * @param user1
+     * @param id
+     * @param userJson
+     * @return
+     * @throws ResourceNotFoundException
+     * @throws InternalErrorException
+     * @throws IOException
+     */
     @Override
     @PUT
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
     @Path("/{id}")
     public User update(@Auth User user1, @PathParam("id") String id, @Valid String userJson) throws ResourceNotFoundException, InternalErrorException, IOException {
-        User user = CommonUtils.convertJsonToObject(userJson, User.class);
-        user.setId(id);
-        try {
-            userDAO.update(getListFromEntity(user));
-            List<User> userList = userDAO.fetchById(getListFromEntityId(id));
-            if (userList != null && !userList.isEmpty()) {
-                return userList.get(0);
+        if (isAdminUser(user1)) {
+            User user = CommonUtils.convertJsonToObject(userJson, User.class);
+            user.setId(id);
+            try {
+                userDAO.update(getListFromEntity(user));
+                List<User> userList = userDAO.fetchById(getListFromEntityId(id));
+                if (userList != null && !userList.isEmpty()) {
+                    return userList.get(0);
+                }
+                throw new ResourceNotFoundException();
+            } catch (Exception e) {
+                throw new InternalErrorException(e);
             }
-            throw new ResourceNotFoundException();
-        } catch (Exception e) {
-            throw new InternalErrorException(e);
+        } else {
+            throw new AuthorizationException("User " + user1.getUserName() + " is not allowed to perform this operation");
         }
     }
 
@@ -115,8 +146,12 @@ public class UserResource extends BaseResource<User> {
     @Path("/{id}")
     public Response delete(@Auth User user, @PathParam("id") String id) throws ResourceNotFoundException, InternalErrorException {
         try {
-            userDAO.remove(getListFromEntityId(id));
-            return Response.ok().build();
+            if (isAdminUser(user)) {
+                userDAO.remove(getListFromEntityId(id));
+                return Response.ok().build();
+            } else {
+                throw new AuthorizationException("User " + user.getUserName() + " is not allowed to perform this operation");
+            }
         } catch (Exception e) {
             throw new ResourceNotFoundException();
         }
